backend/pyproject.toml
[tool.poetry]
name = "bible-search-api"
version = "0.1.0"
description = "AI-powered Bible search API using FastAPI and OpenAI"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app.py"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic = "^2.11.9"
openai = "^1.107.2"
loguru = "^0.7.2"
python-docx = "^1.2.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.3"
reportlab = "^4.4.10"
pillow = "^12.1.1"
pymupdf = "^1.27.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
